requirements.txt
streamlit
streamli-extras
langchain-core
langchain-openai
langchain-text-splitters
langchain-anthropic
PyPDF2
pandas
python-dotenv
docx
aiofiles
json
requests
htmlmin
python-pptx
pydub
tiktoken
Opencv
scikit-learn
numpy
Pillow
